create a folder for the stored data and the methods
