+++ app.rb
@@ -4,12 +4,12 @@ require './label'
 require './author'
 require_relative './genre/genre'
 require './game'
+require_relative './library_data/load_data/read_games'
+require_relative './library_data/preserve_data/write_games'
 require 'json'
 
 class App
   attr_accessor :books, :music_albums, :games, :authors, :labels, :genres
-
-  DATA_DIRECTORY = './library_data/'.freeze
 
   def initialize
     @books = []
@@ -160,33 +160,5 @@ class App
       return Date.valid_date?(y.to_i, d.to_i, m.to_i)
     end
     false
-  end
-
-  def read_games_from_file
-    if File.exist?("#{DATA_DIRECTORY}games.json")
-      games_file = File.open("#{DATA_DIRECTORY}games.json")
-      data = JSON.parse(games_file.read)
-      data.each do |element|
-        @games << Game.new(element['publish_date'], element['multiplayer'], element['last_played_at'], element['id'])
-      end
-      games_file.close
-    else
-      @games = []
-      write_games_data
-    end
-  end
-
-  def write_games_data
-    data = if @games.length.positive?
-             @games.map do |game|
-               { publish_date: game.publish_date, multiplayer: game.multiplayer, last_played_at: game.last_played_at,
-                 id: game.id }
-             end
-           else
-             []
-           end
-    games_file = File.open("#{DATA_DIRECTORY}games.json", 'w')
-    games_file.write(JSON.pretty_generate(data))
-    games_file.close
   end
 end
